consultations/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.12.4
sqlparse==0.4.4
tzdata==2024.1
